Read local text file with all names
Got the script  to read a local text with all of the names (and number of names) and then creates buttons with those names.

Next step is to then read the text files within that person's folder and create buttons with links, etc.

Problem: Need to get this information from a server rather than a local text file. Should just have to change path to localhost (or the IP) and the path to the file(s).